#151 Change according to Luke's review comments
