+++ app/app_pub_test.go
@@ -5,29 +5,18 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	abci "github.com/tendermint/tendermint/abci/types"
+	dbm "github.com/tendermint/tendermint/libs/db"
+	"github.com/tendermint/tendermint/libs/log"
+
 	"github.com/BiJie/BinanceChain/app/config"
 	"github.com/BiJie/BinanceChain/app/pub"
 	"github.com/BiJie/BinanceChain/common/testutils"
 	orderPkg "github.com/BiJie/BinanceChain/plugins/dex/order"
 	dextypes "github.com/BiJie/BinanceChain/plugins/dex/types"
-	"github.com/BiJie/BinanceChain/wire"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/auth"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	abci "github.com/tendermint/tendermint/abci/types"
-	dbm "github.com/tendermint/tendermint/libs/db"
-	"github.com/tendermint/tendermint/libs/log"
-)
-
-var (
-	keeper *orderPkg.Keeper
-	buyer  sdk.AccAddress
-	seller sdk.AccAddress
-	am     auth.AccountMapper
-	ctx    sdk.Context
-	app    *BinanceChain
-	cdc    *wire.Codec
 )
 
 // TODO(#66): fix all time.Sleep - potential source of flaky test
@@ -64,7 +53,7 @@ func TestAppPub_AddOrder(t *testing.T) {
 	assert, require := setupAppTest(t)
 
 	msg := orderPkg.NewNewOrderMsg(buyer, "1", orderPkg.Side.BUY, "XYZ_BNB", 102000, 3000000)
-	keeper.AddOrder(orderPkg.OrderInfo{msg, 100, 0, ""}, 42, false)
+	keeper.AddOrder(orderPkg.OrderInfo{msg, 42, 0, 42, 0, 0, ""}, false)
 	app.EndBlocker(ctx, abci.RequestEndBlock{42})
 	time.Sleep(5 * time.Second)
 
@@ -78,7 +67,7 @@ func TestAppPub_MatchOrder(t *testing.T) {
 	assert, require := setupAppTest(t)
 
 	msg := orderPkg.NewNewOrderMsg(buyer, "1", orderPkg.Side.BUY, "XYZ_BNB", 102000, 3000000)
-	keeper.AddOrder(orderPkg.OrderInfo{msg, 100, 0, ""}, 41, false)
+	keeper.AddOrder(orderPkg.OrderInfo{msg, 41, 100, 41, 100, 0, ""}, false)
 	app.setDeliverState(abci.Header{Height: 41, Time: 100})
 	app.EndBlocker(ctx, abci.RequestEndBlock{41})
 	time.Sleep(5 * time.Second)
@@ -88,7 +77,7 @@ func TestAppPub_MatchOrder(t *testing.T) {
 
 	// we add a sell order to fully execute the buyer order
 	msg = orderPkg.NewNewOrderMsg(seller, "2", orderPkg.Side.SELL, "XYZ_BNB", 102000, 4000000)
-	keeper.AddOrder(orderPkg.OrderInfo{msg, 101, 0, ""}, 42, false)
+	keeper.AddOrder(orderPkg.OrderInfo{msg, 42, 101, 42, 101, 0, ""}, false)
 	app.setDeliverState(abci.Header{Height: 42, Time: 101})
 	app.endBlocker(ctx, abci.RequestEndBlock{42})
 	time.Sleep(5 * time.Second)
@@ -99,9 +88,9 @@ func TestAppPub_MatchOrder(t *testing.T) {
 
 	// we execute qty 1000000 sell order but add a new qty 1000000 sell order, both buy and sell price level should not publish
 	msg = orderPkg.NewNewOrderMsg(buyer, "3", orderPkg.Side.BUY, "XYZ_BNB", 102000, 1000000)
-	keeper.AddOrder(orderPkg.OrderInfo{msg, 102, 0, ""}, 43, false)
+	keeper.AddOrder(orderPkg.OrderInfo{msg, 43, 102, 43, 102, 0, ""}, false)
 	msg = orderPkg.NewNewOrderMsg(seller, "4", orderPkg.Side.SELL, "XYZ_BNB", 102000, 1000000)
-	keeper.AddOrder(orderPkg.OrderInfo{msg, 102, 0, ""}, 43, false)
+	keeper.AddOrder(orderPkg.OrderInfo{msg, 43, 102, 43, 102, 0, ""}, false)
 	app.setDeliverState(abci.Header{Height: 43, Time: 102})
 	app.endBlocker(ctx, abci.RequestEndBlock{43})
 	time.Sleep(5 * time.Second)
